Add context objects to StatisticsCollector methods
The `StatisticsCollector` interface is useful for gathering basic
metrics about data loaders, but is limited in that we are unable to pass
in any sort of context (key or call). This would allow us to obtain more
insightful metrics (for example, we would be able to break down metrics
based on the GraphQL operation name).

To this end, we add `*StatisticsContext` objects to each
`StatisticsCollector` method; we provide distinct implementations for
each one rather than a generic catch-all so that we may evolve them
independently as needed.