Parameterise more tests across DataLoader implementations
We parameterise:

 - `DataLoaderValueCacheTest`
 - `ScheduledDataLoaderRegistryTest`

to provide increased coverage across all DataLoader implementations.

Notably, `DataLoaderValueCacheTest` is _not_ parameterised across the
Publisher DataLoaders (yet) as these do not appear to work when a
`ValueCache` is provided (which is what prompted this pull request). It
will be fixed in a future pull request, at which point we will restore
coverage.